ff-remove-gemini: drop Gemini reviewer support entirely
Operator was being charged for Gemini 2.5 Pro API calls. After Bundle 2
(PR #837), Gemini was the second reviewer at spec and plan only —
tasks/diff/closeout already had no default reviews. Removing Gemini
cleanly leaves Codex-only reviews at spec and plan: ~2 calls per feature.

Implemented by Codex (gpt-5.4-mini, 547k tokens). Discovery-first
approach: Codex grepped every Gemini reference in
docs/workflow/operations/codex-skills/, categorized each as
delete/update/doc-update/leave, then made the changes.

Files deleted (3):
- review-lens/scripts/run_gemini_review.py (the Gemini runner)
- review-lens/scripts/tests/test_run_gemini_review.py
- feature-factory/scripts/tests/fixtures/run-033-state-pre-fix.json
  (dead fixture with Gemini payloads, no test references)

Code removed:
- parse_tokens_gemini + _find_gemini_token_payload from factory_telemetry.py
- gemini_lens branches from required_reviews() in factory_review_specs.py
- DEFAULT_GEMINI_MODEL constant
- --extra-gemini-lens CLI flag and related plumbing
- Gemini reviewer branches in factory_review.py, factory_invariants.py,
  factory_cmd_checkpoint.py, factory_cmd_closeout.py, factory_cmd_deliver.py,
  factory_cmd_dispatch.py, factory_cmd_quick.py, factory_cmd_status.py,
  factory_git.py, factory_state.py, run_factory.py
- Gemini-only assertions across the test suite

Docs updated:
- SKILL.md: stage matrix and review-pipeline language now Codex-only
- CODEX-ORCHESTRATOR.md: same
- feedback.md, pricing.json: Gemini entries removed

Notable side-effect: review-lens test suite is now empty (its only test
file was the Gemini one). FF suite: 288 tests passing.

Net: 40 files changed, +156 / -3019 (net -2,863 lines). If Gemini
support is ever wanted back, restore from git history.
