+++ docs/workflow/operations/codex-skills/feature-factory/scripts/factory_cmd_deliver.py
@@ -538,7 +538,6 @@ def command_workflow_closeout(args: argparse.Namespace) -> int:
         base_ref=None,
         context=[],
         path=[],
-        extra_gemini_lens=[],
         sensitive=False,
         large_structural=False,
         performance_sensitive=False,
@@ -547,8 +546,6 @@ def command_workflow_closeout(args: argparse.Namespace) -> int:
         max_artifact_chars=None,
         max_context_chars=None,
         max_total_chars=None,
-        gemini_timeout_seconds=120,
-        gemini_retries=1,
         repair_timeout_seconds=300,
         fallback=False,
         allow_large_diff_rerun=False,
