+++ docs/workflow/operations/codex-skills/feature-factory/scripts/factory_review.py
@@ -48,9 +48,7 @@ from factory_stages import (  # noqa: E402
 from factory_next_action import recommended_next_action  # noqa: E402
 
 from factory_review_specs import (  # noqa: E402,F401
-    DEFAULT_GEMINI_MODEL,
     DEFAULT_CODEX_MODEL,
-    SMALL_TASK_SET_THRESHOLD,
     _AUTO_ACCEPT_NOTE,
     detect_actionable_findings,
     trim_detail,
@@ -66,14 +64,7 @@ WRITE_DIFF = REVIEW_SCRIPTS / "write_canonical_diff.py"
 REPAIR = REVIEW_SCRIPTS / "repair_review_checkpoint.py"
 UPDATE_REVIEW = REVIEW_SCRIPTS / "update_review_resolution.py"
 APPEND_RECONCILIATION = REVIEW_SCRIPTS / "append_reconciliation_entry.py"
-RUN_GEMINI_REVIEW = REVIEW_SCRIPTS / "run_gemini_review.py"
 RUN_CODEX_REVIEW = REVIEW_SCRIPTS / "run_codex_review.py"
-
-# Gemini reviews launch GEMINI_STAGGER_SECONDS apart without the file lock so they overlap.
-# Codex reviews always run fully in parallel with Gemini (different API, no rate limit concern).
-# Validated 2026-03-30: zero 429s across parallel-implement-command checkpoints at 30s stagger.
-# Set to None to revert to strict serial execution if rate-limit errors reappear.
-GEMINI_STAGGER_SECONDS: int | None = 30
 
 # ---------------------------------------------------------------------------
 # Auto-context file extraction (used by command_checkpoint in factory_cmd_checkpoint)
@@ -127,7 +118,6 @@ def resolved_review_policy(slug: str, args: argparse.Namespace) -> dict:
             "sensitive": False,
             "large_structural": False,
             "performance_sensitive": False,
-            "extra_gemini_lenses": [],
         },
     )
     if args.sensitive:
@@ -136,8 +126,6 @@ def resolved_review_policy(slug: str, args: argparse.Namespace) -> dict:
         policy["large_structural"] = True
     if args.performance_sensitive:
         policy["performance_sensitive"] = True
-    if args.extra_gemini_lens:
-        policy["extra_gemini_lenses"] = list(args.extra_gemini_lens)
     save_workflow_state(slug, state)
     return policy
 
@@ -215,7 +203,6 @@ def repair_checkpoint_args(slug: str, stage: str, state: dict[str, object]) -> a
         "context": context_paths,
         "path": [],
         "required_reviews": manifest.get("required_reviews"),
-        "extra_gemini_lens": [],
         "sensitive": False,
         "large_structural": False,
         "performance_sensitive": False,
@@ -224,8 +211,6 @@ def repair_checkpoint_args(slug: str, stage: str, state: dict[str, object]) -> a
         "max_artifact_chars": manifest.get("max_artifact_chars"),
         "max_context_chars": manifest.get("max_context_chars"),
         "max_total_chars": manifest.get("max_total_chars"),
-        "gemini_timeout_seconds": 120,
-        "gemini_retries": 1,
         "repair_timeout_seconds": 300,
         "allow_large_diff_rerun": bool(diff_budget.get("artifact_changed_since_codex")),
         "fallback": False,
@@ -243,10 +228,10 @@ def repair_checkpoint_args(slug: str, stage: str, state: dict[str, object]) -> a
 # ---------------------------------------------------------------------------
 
 
-def fallback_review_command(spec: dict, artifact_path: Path, checkpoint: dict, workspace_dir: Path | None, timeout_seconds: int, retries: int) -> list[str]:
+def fallback_review_command(spec: dict, artifact_path: Path, checkpoint: dict, workspace_dir: Path | None) -> list[str]:
     cmd = [
         sys.executable,
-        str(RUN_GEMINI_REVIEW if spec.get("reviewer") == "gemini" else RUN_CODEX_REVIEW),
+        str(RUN_CODEX_REVIEW),
         "--artifact",
         str(artifact_path),
         "--lens",
@@ -270,12 +255,10 @@ def fallback_review_command(spec: dict, artifact_path: Path, checkpoint: dict, w
         cmd.extend(["--max-total-chars", str(checkpoint["max_total_chars"])])
     if workspace_dir:
         cmd.extend(["--workspace-dir", str(workspace_dir)])
-    if spec.get("reviewer") == "gemini":
-        cmd.extend(["--timeout-seconds", str(timeout_seconds), "--retries", str(retries)])
     return cmd
 
 
-def run_checkpoint_fallback(manifest_path: Path, workspace_root: Path, gemini_timeout_seconds: int, gemini_retries: int) -> tuple[bool, str]:
+def run_checkpoint_fallback(manifest_path: Path, workspace_root: Path) -> tuple[bool, str]:
     checkpoint = json.loads(read_text(manifest_path))
     artifact_path = resolve_stored_path(checkpoint["artifact_path"], REPO_ROOT)
     workspace_dir = workspace_root.resolve()
@@ -290,45 +273,22 @@ def run_checkpoint_fallback(manifest_path: Path, workspace_root: Path, gemini_ti
             data = {}
         return artifact_hash_matches(checkpoint.get("stage", ""), artifact_path, data)
 
-    def _run_review(spec: dict, no_gemini_lock: bool = False) -> str | None:
+    def _run_review(spec: dict) -> str | None:
         if _already_done(spec):
             return None
-        cmd = fallback_review_command(spec, artifact_path, checkpoint, workspace_dir, gemini_timeout_seconds, gemini_retries)
-        if no_gemini_lock and spec.get("reviewer") == "gemini":
-            cmd.append("--no-gemini-lock")
-        timeout = gemini_timeout_seconds + 30 if spec.get("reviewer") == "gemini" else 210
-        result = subprocess.run(cmd, text=True, capture_output=True, timeout=timeout)
+        cmd = fallback_review_command(spec, artifact_path, checkpoint, workspace_dir)
+        result = subprocess.run(cmd, text=True, capture_output=True, timeout=210)
         if result.returncode != 0:
             return f"{spec['path']}: {trim_detail(result.stderr or result.stdout or 'review failed')}"
         return None
 
     all_specs = checkpoint.get("required_reviews", [])
-    gemini_specs = [s for s in all_specs if s.get("reviewer") == "gemini"]
-    codex_specs = [s for s in all_specs if s.get("reviewer") != "gemini"]
-    parallel_gemini = GEMINI_STAGGER_SECONDS is not None
-
     failed: list[str] = []
     futures: list[concurrent.futures.Future[str | None]] = []
 
     with concurrent.futures.ThreadPoolExecutor() as executor:
-        # Codex always runs in parallel with Gemini (different API, no rate limit concern).
-        for spec in codex_specs:
+        for spec in all_specs:
             futures.append(executor.submit(_run_review, spec))
-
-        if parallel_gemini:
-            # EXPERIMENT: launch Gemini reviews GEMINI_STAGGER_SECONDS apart without the
-            # file lock. See constant definition above for tuning / rollback instructions.
-            for i, spec in enumerate(gemini_specs):
-                if i > 0:
-                    time.sleep(GEMINI_STAGGER_SECONDS)
-                futures.append(executor.submit(_run_review, spec, True))
-        else:
-            # Serial fallback: run Gemini reviews one at a time (rate-limit safe).
-            for spec in gemini_specs:
-                err = _run_review(spec)
-                if err:
-                    failed.append(err)
-
         for future in concurrent.futures.as_completed(futures):
             err = future.result()
             if err:
