+++ docs/workflow/operations/codex-skills/feature-factory/scripts/tests/test_factory_cmd_checkpoint.py
@@ -79,7 +79,6 @@ def _args(json_flag: bool = False) -> argparse.Namespace:
         base_ref=None,
         context=[],
         path=[],
-        extra_gemini_lens=[],
         sensitive=False,
         large_structural=False,
         performance_sensitive=False,
@@ -89,8 +88,6 @@ def _args(json_flag: bool = False) -> argparse.Namespace:
         max_artifact_chars=50000,
         max_context_chars=100000,
         max_total_chars=180000,
-        gemini_timeout_seconds=120,
-        gemini_retries=1,
         repair_timeout_seconds=300,
         allow_large_diff_rerun=False,
         fallback=False,
