+++ docs/workflow/operations/codex-skills/feature-factory/scripts/tests/test_review_gc.py
@@ -62,7 +62,6 @@ def _checkpoint_args(*, keep_intermediates: bool = False) -> argparse.Namespace:
         base_ref=None,
         context=[],
         path=[],
-        extra_gemini_lens=[],
         sensitive=False,
         large_structural=False,
         performance_sensitive=False,
@@ -72,8 +71,6 @@ def _checkpoint_args(*, keep_intermediates: bool = False) -> argparse.Namespace:
         max_artifact_chars=50000,
         max_context_chars=100000,
         max_total_chars=180000,
-        gemini_timeout_seconds=120,
-        gemini_retries=1,
         repair_timeout_seconds=300,
         allow_large_diff_rerun=False,
         fallback=False,
