+++ docs/workflow/operations/codex-skills/feature-factory/scripts/tests/test_self_documenting_errors.py
@@ -72,7 +72,6 @@ class SelfDocumentingErrorTests(unittest.TestCase):
             "base_ref": None,
             "context": [],
             "path": [],
-            "extra_gemini_lens": [],
             "sensitive": False,
             "large_structural": False,
             "performance_sensitive": False,
@@ -82,8 +81,6 @@ class SelfDocumentingErrorTests(unittest.TestCase):
             "max_artifact_chars": 50000,
             "max_context_chars": 60000,
             "max_total_chars": 250000,
-            "gemini_timeout_seconds": 120,
-            "gemini_retries": 1,
             "repair_timeout_seconds": 300,
             "allow_large_diff_rerun": False,
             "fallback": False,
